pyproject.toml
[tool.poetry]
name = "bash_command_finder"
version = "0.1.0"
description = ""
authors = ["Sam Purkis <sam.purkis@hotmail.co.uk>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
requests = "^2.28.1"
click = "^8.1.3"
colorama = "^0.4.5"
pyperclip = "^1.8.2"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^5.3"
nuitka = "^1.4.1"
zstandard = "^0.19.0"
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[mypy]
warn_unused_configs = true
